+++ neo/fixtures_test.go
@@ -72,10 +72,10 @@ var (
 	}{
 		{
 			asset: "0xc56f33fc6ecfcd0c225c4ab356fee59390af8560be0e930faebe74a6daff7c9b",
-			hash:  "0xc515c4d2db27e06fd2305a5c5378f820d2c4cc04477ebe40ffa40b956eb4f8b5",
+			hash:  "0xd30e2fe0bc3ecb1fde65f611a7f827268535ab19d2d2647d9984f35719185945",
 			id:    "1",
 			index: 0,
-			value: "96",
+			value: "6",
 		},
 	}